Like button is working | Bookmarks is working | Download button is working

+++ app/src/main/java/in/cipherhub/notebox/BookmarkActivity.java
@@ -127,7 +127,7 @@ public class BookmarkActivity extends AppCompatActivity {
                 , pdf.getString("date")
                 , pdf.getInt("shares")
                 , pdf.getInt("downloads")
-                , pdf.getDouble("rating")
+                , pdf.getInt("rating")
         ));
       } catch (JSONException e) {
         Log.d(TAG, String.valueOf(e));
